+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All substantial Mapwatch website changes.
 
 ---
 
+## 2021-03-13
+- Syndicate and Mastermind hideouts are now tracked.
+- The Domain of Timeless Conflict is now tracked.
+
 ## 2021-03-09
 - Enabled exporting your map data to Google Sheets. It's on the history screen, next to the TSV-export button.